Revert Move module
