+++ OTAnalytics/plugin_ui/customtkinter_gui/toplevel_flows.py
@@ -5,7 +5,7 @@ from typing import Any, Optional
 from customtkinter import CTkEntry, CTkLabel, CTkOptionMenu
 
 from OTAnalytics.application.application import CancelAddFlow
-from OTAnalytics.application.generate_flows import FlowNameGenerator
+from OTAnalytics.application.use_cases.generate_flows import FlowNameGenerator
 from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY
 from OTAnalytics.plugin_ui.customtkinter_gui.toplevel_template import (
     FrameContent,
