+++ tests/OTAnalytics/application/use_cases/test_generate_flows.py
@@ -2,8 +2,7 @@ from itertools import chain, repeat
 from unittest.mock import Mock, call
 
 import pytest
-
-from OTAnalytics.application.generate_flows import (
+from application.use_cases.generate_flows import (
     CrossProductFlowGenerator,
     FilterSameSection,
     FlowGenerator,
@@ -12,6 +11,7 @@ from OTAnalytics.application.generate_flows import (
     FlowPredicate,
     GenerateFlows,
 )
+
 from OTAnalytics.domain.flow import Flow, FlowId, FlowRepository
 from OTAnalytics.domain.section import Section, SectionId, SectionRepository
